+++ client/src/app/(pages)/(home)/page.module.scss
@@ -1,1 +1,29 @@
-@use "app/(pages)/_globals/mixins.scss";
+@use "../_globals/mixins";
+
+.pageWrapper {
+  padding: 0;
+  margin: 0;
+}
+
+.searchHeading {
+  position: absolute;
+  top: 144px;
+  left: 108px;
+  width: 203px;
+  height: 27px;
+  margin: 0;
+}
+
+.contentRow {
+  position: absolute;
+  top: 239px;
+  left: 103px;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 24px; 
+}
+
+.listingsArea {
+  flex: 1;
+}
