+++ client/src/app/(pages)/_components/HomepageFilterMenu/HomepageFilterMenu.tsx
@@ -1,7 +1,7 @@
 "use client";
 import styles from "./HomepageFilterMenu.module.scss";
 import React from "react";
-import FilterDropdown from "./FilterDropdown";
+import FilterDropdown from "../FilterDropdown/FilterDropdown";
 
 const DROPDOWN_OPTIONS: { label: string; options: string[] }[] = [
   {
@@ -10,11 +10,22 @@ const DROPDOWN_OPTIONS: { label: string; options: string[] }[] = [
   },
   {
     label: "Price Range",
-    options: ["Under $250", "$250 - $500", "$500 - $750", "$750 - $1000", "$1000+"],
+    options: [
+      "Under $250",
+      "$250 - $500",
+      "$500 - $750",
+      "$750 - $1000",
+      "$1000+",
+    ],
   },
   {
     label: "Sort By",
-    options: ["Relevance", "Newly Listed", "Price: Low to High", "Price: High to Low"],
+    options: [
+      "Relevance",
+      "Newly Listed",
+      "Price: Low to High",
+      "Price: High to Low",
+    ],
   },
 ];
 
@@ -26,7 +37,7 @@ export default function HomepageFilterMenu() {
       <div className={styles.dropdownGroup}>
         {DROPDOWN_OPTIONS.map(({ label, options }) => (
           <FilterDropdown key={label} label={label} options={options} />
-        ))}        
+        ))}
       </div>
 
       <button className={styles.applyButton}>Apply</button>
